requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiosqlite==0.19.0
sqlalchemy==2.0.20
torch==2.1.0
transformers==4.33.3
sentencepiece==0.1.99
numpy==1.24.3
scikit-learn==1.3.0
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.0.4
tqdm==4.65.0
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.9.1
mypy==1.5.1
types-requests==2.31.0.1
pylint==2.17.4